requirements.txt
SpeechRecognition
python-dotenv
langchain
langchain-community
langchain_huggingface
Huggingface-hub
Pyaudio
keyboard
pyttsx3